refactor: move lab_manager access check to require_access() header
